+++ reactjs/src/components/molecules/challenges/ChallengeBody.jsx
