+++ reactjs/src/components/molecules/challenges/ChallengeNavbar.jsx
