+++ reactjs/src/components/molecules/challenges/ChallengeOverview.jsx
